fixing circular import and mpmathifying constants
