+++ precise.py
@@ -1,60 +1,12 @@
 import pvlib
 import numpy as np
 import matplotlib.pyplot as plt
-from mpmath import mp
 import itertools
 import json
-import csv
 import datetime
 import max_power
-
-def diff_lhs_rhs(v, i, il, io, rs, rsh, n, vth, ns):
-    r"""
-    Calculates the difference between the left hand side and right hand side of
-    the single diode equation.
-
-    Parameters
-    ----------
-    v : numeric
-        Voltage [V]
-
-    i : numeric
-        Current [A]
-
-    il : numeric
-        Light-generated current :math:`I_L` (photocurrent) [A]
-
-    io : numeric
-        Diode saturation :math:`I_0` current under desired IV curve conditions.
-        [A]
-
-    rs : numeric
-        Series resistance :math:`R_s` under desired IV curve conditions. [ohm]
-
-    rsh : numeric
-        Shunt resistance :math:`R_{sh}` under desired IV curve conditions.
-        [ohm]
-
-    n : numeric
-        Diode ideality factor :math:`n`
-
-    vth : numeric
-        Thermal voltage of the cell :math:`V_{th}` [V]
-        The thermal voltage of the cell (in volts) may be calculated as
-        :math:`k_B T_c / q`, where :math:`k_B` is Boltzmann's constant (J/K),
-        :math:`T_c` is the temperature of the p-n junction in Kelvin, and
-        :math:`q` is the charge of an electron (coulombs). 
-
-    ns : numeric
-        Number of cells in series :math:`N_s`
-
-    Returns
-    -------
-    mpmath float
-        Difference between the left hand and right hand sides of the single
-        diode equation.
-    """
-    return (il - io*mp.expm1((v + i*rs)/(n*ns*vth)) - (v + i*rs)/(rsh) - i)
+import utils
+from utils import mp, diff_lhs_rhs
 
 
 def get_precise_i(il, io, rs, rsh, n, vth, ns, atol, num_pts):
@@ -171,22 +123,6 @@ def plot_case_iv_curves(case_title, case_parameter_sets, vth, atol, num_pts):
     plt.show()
 
 
-def read_case_parameters(filename):
-    with open(f'{filename}.csv', newline='') as file:
-        reader = csv.DictReader(file, delimiter=',')
-        parameter_columns = ['photocurrent', 'saturation_current',
-                             'resistance_series', 'resistance_shunt', 'n',
-                             'cells_in_series']
-        rows = []
-        for row in reader:
-            rows.append(
-                [int(row['Index'])]
-                + [mp.mpmathify(row[col], strings=True)
-                    for col in parameter_columns]
-            )
-        return rows
-
-
 def write_case_tests(case_filename, case_parameter_sets, vth, temp_cell, atol,
                      num_pts):
     case_test_suite = {'Manufacturer': '', 'Sandia ID': '', 'Material': '',
@@ -224,18 +160,14 @@ def write_case_tests(case_filename, case_parameter_sets, vth, temp_cell, atol,
 
 
 if __name__ == "__main__":
-    mp.dps = 40 # set precision, 16*2 rounded up
-    num_pts = 100
-    atol = mp.mpmathify(1e-16)
-
-    # Boltzmann's const (J/K), electron charge (C), temp (K) 
-    k, q, temp_cell = map(mp.mpmathify, [1.380649e-23, 1.60217663e-19, 298.15])
-    vth = (k * temp_cell) / q
-
     case_number = 1
     case_filename = f'tests/case{case_number}'
     case_title = f'Case {case_number}'
-    case_parameter_sets = read_case_parameters(case_filename)
+
+    constants = utils.constants()
+    vth, temp_cell, atol, num_pts = (constants['vth'], constants['temp_cell'],
+                                     constants['atol'], constants['num_pts'])
+    case_parameter_sets = utils.read_case_parameters(case_filename)
 
     write_case_tests(case_filename, case_parameter_sets, vth, temp_cell, atol,
                      num_pts)
